Move inspec init to v2 plugins.
