+++ lib/plugins/inspec-init/lib/inspec-init/templates/profile/README.md
