+++ lib/plugins/inspec-init/lib/inspec-init/templates/profile/controls/example.rb
