+++ lib/plugins/inspec-init/lib/inspec-init/templates/profile/inspec.yml
